+++ api/acceptance-test/src/test/java/me/nalab/luffy/api/acceptance/test/feedback/FeedbackAcceptanceValidator.java
@@ -13,9 +13,7 @@ public final class FeedbackAcceptanceValidator {
 		throw new UnsupportedOperationException("Cannot invoke constructor \"FeedbackAcceptanceValidator()\"");
 	}
 
-	//:TODO 여기에 Feedback 인수테스트 validator 로직을 static으로 작성 해주세요. SurveyAcceptanceValidator 참고해서 작성하면 됩니다.
 
-	//: TODO
 	public static void assertIsFeedbackSummaryFound(ResultActions resultActions) throws Exception {
 		resultActions.andExpectAll(
 			status().isOk(),
